requirements.txt
langchain
langchain_groq
streamlit
deepgram-sdk
langchain_community
python-dotenv